separate CarController and ProductController to different files
